Move setup-related files to separate folder
