+++ Source/SelfService/Web/integrations/connection/connectionDetails/configuration/setup/ActionToolbar.tsx
