+++ Source/SelfService/Web/integrations/connection/connectionDetails/configuration/setup/ConfigurationFormContent.tsx
