+++ Source/SelfService/Web/integrations/connection/connectionDetails/configuration/setup/DeleteConnectorDialog.tsx
