+++ Source/SelfService/Web/integrations/connection/connectionDetails/configuration/setup/M3AuthenticationType.ts
