+++ Source/SelfService/Web/integrations/connection/connectionDetails/configuration/setup/MetadataPublisherCredentials.tsx
